🎨 clean up our components some more
